GXBeamCS

Documentation for GXBeamCS.

GXBeamCS.FEM

GXBeamCS.Layer

GXBeamCS.Material

GXBeamCS.MeshElement

GXBeamCS.Node

GXBeamCS.SectionCache

GXBeamCS.MaterialPlane

GXBeamCS.Qbar

GXBeamCS.addwebs

GXBeamCS.afmesh

GXBeamCS.area_and_centroid_of_element

GXBeamCS.clt_Talpha

GXBeamCS.clt_Ttheta

GXBeamCS.combine_halfs

GXBeamCS.compliance_matrix

GXBeamCS.compliance_matrix

GXBeamCS.elementQ

GXBeamCS.element_integrand

GXBeamCS.element_orientation

GXBeamCS.element_submatrix

GXBeamCS.find_inner_surface

GXBeamCS.initialize_cache

GXBeamCS.insertpoint

GXBeamCS.laminatecompliance

GXBeamCS.laminatecompliancematrix

GXBeamCS.laminatestiffnessmatrix

GXBeamCS.linearsolve

GXBeamCS.mass_matrix

GXBeamCS.mass_matrix

GXBeamCS.mesh_cylinder

GXBeamCS.node2idx

GXBeamCS.node2idx!

GXBeamCS.nodes_half

GXBeamCS.parse_BECAS

GXBeamCS.parseairfoil

GXBeamCS.plotgeometry

GXBeamCS.plotgeometry

GXBeamCS.plotsoln

GXBeamCS.plotsoln

GXBeamCS.preprocess_layers

GXBeamCS.redistribute_thickness

GXBeamCS.reorder

GXBeamCS.resample

GXBeamCS.rotate_element_to_beam

GXBeamCS.rotate_ply_to_element

GXBeamCS.rotate_stress_and_strains

GXBeamCS.stiffness

GXBeamCS.strains

GXBeamCS.strains_and_stresses

GXBeamCS.strains_and_stresses

GXBeamCS.stresses

GXBeamCS.tangential

GXBeamCS.te_inner_intersection

GXBeamCS.tsai_wu

GXBeamCS.tsai_wu

GXBeamCS.web_intersections

GXBeamCS.zspacing

GXBeamCS.zspacingdouble_fromz

GXBeamCS.FEM — Type

FEM(nodes, elements; cache=initialize_cache(nodes, elements))

Initialize finite element object.

Arguments

nodes: Vector containing all the nodes in the mesh

elements: Vector containing all the elements in the mesh

cache::SectionCache: A pre-allocated cache which may be passed in to reduce allocations across multiple calls to this function when the number of nodes, number of elements, and connectivity remain the same.

GXBeamCS.Layer — Type

Layer(material, t, theta)

A layer (could be one ply or many plys of same material). A layup is a vector of layers.

Arguments

material::Material: material of layer

t::float: thickness of layer

theta::float: fiber orientation (rad)

GXBeamCS.Material — Type

Material(E1, E2, E3, G12, G13, G23, nu12, nu13, nu23, rho,
    S1t=1.0, S1c=1.0, S2t=1.0, S2c=1.0, S3t=1.0, S3c=1.0, S12=1.0, S13=1.0, S23=1.0)

Orthotropic material properties. Axis 1 is the main ply axis, axis 2 is the transverse ply axis, and axis 3 is normal to the ply. For a fiber orientation of zero, axis 1 is along the beam axis.

Arguments

Ei::float: Young's modulus along ith axis.

Gij::float: Shear moduli

nuij::float: Poisson's ratio. $nu_ij E_j = nu_ji E_i$

rho::float: Density

Sit::float: Tensile strength in ith direction

Sic::float: Compressive strength in ith direction

Sij::float: Strength in ij direction

GXBeamCS.MeshElement — Type

MeshElement(nodenum, material, theta)

An element in the mesh, consisting of four ordered nodes, a material, and a fiber orientation.

Arguments

nodenum::Vector{integer}: Node indices, defined counterclockwise starting from the bottom left node (as defined in the local coordinate system, see the figure).

material::Material: Material properties

theta::float: Fiber orientation

GXBeamCS.Node — Type

Node(x, y)

A node in the finite element mesh at location x, y. If assembled in a vector, the vector index is the node number.

Arguments

x::float: x-location of node in global coordinate system

y::float: y-location of node in global coordinate system

GXBeamCS.SectionCache — Type

SectionCache{TM, TSM, TMF, TSMF, TAF, TSAF}

Internal cache for section properties

GXBeamCS.MaterialPlane — Method

MaterialPlane(E1, E2, G12, nu12, rho,
    S1t=1.0, S1c=1.0, S2t=1.0, S2c=1.0, S12=1.0)

Orthotropic material properties for a plane stress case (which assumes no stress along axis 3). Axis 1 is the main ply axis, axis 2 is the transverse ply axis, and axis 3 is normal to the ply. For a fiber orientation of zero, axis 1 is along the beam axis.

Arguments

Ei::float: Young's modulus along ith axis.

Gij::float: Shear moduli

nuij::float: Poisson's ratio. $nu_ij E_j = nu_ji E_i$

rho::float: Density

Sit::float: Tensile strength in ith direction

Sic::float: Compressive strength in ith direction

Sij::float: Strength in ij direction

GXBeamCS.Qbar — Method

Qbar(lamina)

Computes the lamina stiffness matrix at some arbitrary orientation theta Transforms a specially orthotropic lamina from principal axis to an arbitrary axis defined by the ply orientation.

GXBeamCS.addwebs — Function

add the nodes and elements for the webs. given the x locations (by idx) where the webs start and the number of grid points in the webs.

GXBeamCS.afmesh — Method

afmesh(xaf, yaf, chord, twist, paxis, xbreak, webloc, segments, webs; ds=nothing, dt=nothing, ns=nothing, nt=nothing, wns=4, wnt=nothing)

Create structural mesh for airfoil. The airfoil coordinates define the meshing density tangential to the airfoil. Whereas the number of layers defines the resolution normal to the airfoil. All segments are meshed with the same resolution in the normal direction, using the number of grid points as defined by segment with the most layers.

Arguments

xaf, yaf::Vector{float}: points defining airfoil start at trailing edge and traverse counterclockwise back to trailing edge (can be blunt or sharp T.E.)

chord::float: chord length

twist::float: twist angle (rad)

paxis::float: pitch axis (normalized by chord). e.g., 0.25 means twist about quarter chord.

xbreak::Vector{float}: x-locations, normalized by chord, defining break points between segments. must start at 0 and end at 1. e.g., [0, 0.2, 0.4, 0.7, 1.0] defines 4 segments.

webloc::Vector{float}: x-locations, normalized by chord, defining web centers (and length of vector is number of webs). e.g., [0.25, 0.55] means there is a web at 25% chord and a second web at 55% chord.

segments::Vector{Vector{Layer}}: A Layer defines a ply (or multiple plys with the same material/orientation). At a given x location the ply stack (segment) is defined a vector of layers starting from outside surface towards inside. Segments then is a vector of these segments that defines properties between segments as defined by xbreak.

webs::Vector{Vector{Layer}}: same structure as segments, except each inner vector is from left to right (although this is usually symmetric), and each outer vector is for a separate web

ds::float: if provided, airfoil spacing will be resampling with approximately this spacing, normalized by chord. e.g., 0.01 will have points on the airfoil roughly 1% chord apart.

dt::float: if provided, thickness will be resampled with this maximum mesh size (thickness is absolute). Note that the total number of cells remains constant along airfoil, so most thicknesses will be much less. e.g., 0.01 will target a maximum mesh thickness of 0.01 (absolute).

ns::vector{int}: if provided, rather than use a targert size ds, we specify the number of cells to use in each segment. This is desirable for gradient-based optimization, if airfoil coordinates are changed, so that during resizing operations the mesh stretch/shrinks rather than experiencing discrete jumps. For example, ns=[15, 20, 40, 30] would use 15 elements between xbreak[1] and xbreak[2] and so on.

nt::vector{vector{int}}: if provided, defines how many elements to use across tangential direction. Again, prefered over dt for gradient-based optimization, if the thicknesses are changed during optimization. each entry defines how many cells to put in that layer following order of original layup. for example, nt=[[1, 2, 1], [1, 3]] would use 1 element, 2 elements (subdivide), then 1 elements over first sector, and so on.

wns::int: discretization level for number of elements vertically along web.

wnt::vector{vector{int}}: same definition as nt but for the webs

Returns

nodes::Vector{Node{Float64}}: nodes for this mesh

elements::Vector{MeshElement{Float64}}: elements for this mesh

GXBeamCS.area_and_centroid_of_element — Method

area_and_centroid_of_element(node)

Compute area and centroid of an element specified by its four nodes

GXBeamCS.clt_Talpha — Method

from primed to unprimed

GXBeamCS.clt_Ttheta — Method

from double prime (e1, e2, e3) to primed c.s.

GXBeamCS.combine_halfs — Method

given nodes/elements for upper surface and lower surface separately, combine into one set making sure to reuse the common nodes that occur at the LE/TE

GXBeamCS.compliance_matrix — Function

compliance_matrix(fem::FEM, shear_center=true)

Compute compliance matrix given a finite element mesh described by nodes and elements.

Arguments

fem::FEM: finite element description (nodes, elements, and cache)

shear_center::Bool: Indicates whether the compliance matrix should be provided about the shear center

Returns

S::Matrix: compliance matrix

sc::Vector{float}: x, y location of shear center (location where a transverse/shear force will not produce any torsion, i.e., beam will not twist)

tc::Vector{float}: x, y location of tension center, aka elastic center, aka centroid (location where an axial force will not produce any bending, i.e., beam will remain straight)

GXBeamCS.compliance_matrix — Function

compliance_matrix(::CompositeSectionAnalysis, shear_center=true)

Compute compliance matrix for the section.

Arguments

shear_center::Bool: Indicates whether the compliance matrix should be provided about the shear center

Returns

S::Matrix: compliance matrix in the order expected by GXBeam.

sc::Vector{float}: x, y location of shear center (location where a transverse/shear force will not produce any torsion, i.e., beam will not twist)

tc::Vector{float}: x, y location of tension center, aka elastic center, aka centroid (location where an axial force will not produce any bending, i.e., beam will remain straight)

GXBeamCS.elementQ — Method

elementQ(material, theta, cbeta, sbeta)

Calculate element constitutive matrix Q accounting for fiber orientation theta and element orientation beta (where cbeta = cos(beta) and sbeta=sin(beta))

GXBeamCS.element_integrand — Method

element_integrand(ksi, eta, element, nodes)

Compute the integrand for a single element with a given ksi, eta.

GXBeamCS.element_orientation — Method

element_orientation(nodes)

Calculate the orientation beta of a single element using its nodes. Return cos(beta) and sin(beta)

GXBeamCS.element_submatrix — Method

element_submatrix(element, nodes)

Performs 2-point Gauss quadrature of one element

GXBeamCS.find_inner_surface — Method

determine curve for inner surface to find intersections at trailing edge and with webs

GXBeamCS.initialize_cache — Method

initialize_cache([TF,] nodes, elements)

Construct the cache for the section property calculations.

GXBeamCS.insertpoint — Method

add new point xn into existing airfoil coordinates x, y. Used to ensure breakpoints are in airfoil mesh. assumes x is already sorted (and thus represents either upper or lower half)

GXBeamCS.laminatecompliance — Method

convenicne method to go from laminate definition to its compliance matrices

GXBeamCS.laminatecompliancematrix — Method

Compute the alpha, beta, delta compliance matrices for a thin laminate given the stiffness matrices

GXBeamCS.laminatestiffnessmatrix — Method

Compute the A, B, D stiffness matrices for a thin laminate

GXBeamCS.linearsolve — Method

linearsolve(A, B; kwargs...)

Linear solve which is overloaded with analytic derivatives. Returns the factorized matrix Afact and the result of the linear solve X.

GXBeamCS.mass_matrix — Method

mass_matrix(fem::FEM)

Compute mass matrix for the section

Returns

M::Matrix: mass matrix in the order expected by GXBeam.

mc::Vector{float}: x, y location of mass center

GXBeamCS.mass_matrix — Method

mass_matrix(::CompositeSectionAnalysis)

Compute mass matrix for the section

Returns

M::Matrix: mass matrix in the order expected by GXBeam.

mc::Vector{float}: x, y location of mass center

GXBeamCS.mesh_cylinder — Method

mesh_cylinder(R, thickness, material)

A simple function to mesh a circular cross-section.

Inputs

r::Vector{Float} - A vector of all the radial location of nodes. Assumed from least to greatest.

materials::Vector{Material} - All of the materials used in the cross section.

material_idx::Vector{Int} - The index of the material in the materials vector for each layer.

Assumed in order of outermost layer to inner most layer. Note: This is opposite to the order of r.

plyangles::Vector{Float} - The ply angle of a given radial element (same order as material_idx).

nt::Int - Number of tangential points (How many elements around the circle).

GXBeamCS.node2idx! — Method

node2idx!(idx, nodenums)

In-place version of node2idx.

GXBeamCS.node2idx — Method

node2idx(nodenums)

Convenience function to map node numbers to locations in global matrix.

GXBeamCS.nodes_half — Method

create nodes and elements for half (upper or lower) portion of airfoil

GXBeamCS.parse_BECAS — Method

parse_BECAS(path)

Convert BECAS input files (N2D.in, E2D.in, EMAT.in, MATPROPS.in) into a vector of GXBeam.GXBeamCS.Node and a vector GXBeam.GXBeamCS.MeshElement.

Arguments

path::String - The path to the BECAS input files.

GXBeamCS.parseairfoil — Function

airfoil coordinates xaf, yaf are assumed to start at trailing edge and traverse counterclockwise back to trailing edge. (thus upper surface first) coordinates are also assumed to be chord-normalized so that xaf is between 0 and 1. xbreak is a vector defining locations that need to be inserted into airfoil coordinates (for upper and lower surfaces) ds is a desired spacing for resampling (set ds=nothing to not resample) Alternatively can use vector nseg to define number of points in each interval between xbreak (see resample())

Method separates airfoil into upper and lower surfaces, adds in break points, and optionally resamples.

GXBeamCS.plotgeometry — Method

plotgeometry(fem::FEM, pyplot; plotnumbers=false)

plot nodes and elements for a quick visualization. Need to pass in a PyPlot object as PyPlot is not loaded by this package.

GXBeamCS.plotgeometry — Method

plotgeometry(::CompositeSectionAnalysis, pyplot; plotnumbers=false)

plot geometry for a quick visualization. Need to pass in a PyPlot object as PyPlot is not loaded by this package.

GXBeamCS.plotsoln — Method

plotsoln(fem::FEM, soln, pyplot)

plot stress/strain on mesh soln could be any vector that is of length # of elements, e.g., sigma_b[3, :] Need to pass in a PyPlot object as PyPlot is not loaded by this package.

GXBeamCS.plotsoln — Method

plotsoln(::CompositeSectionAnalysis, soln, pyplot)

plot stress/strain on mesh soln could be any vector that is of appropriate length for the analysis method, e.g., sigma_b[3, :] Need to pass in a PyPlot object as PyPlot is not loaded by this package.

GXBeamCS.preprocess_layers — Function

convert segments to all have the same number of layers for ease in meshing. Overall definition remains consistent, just break up some thicker layers into multiple thinner layers (with same material and orientation properties)

GXBeamCS.redistribute_thickness — Method

Modify number of layers based on a given maximum thickness

GXBeamCS.reorder — Method

reorder(K)

Reorder stiffness or compliance matrix from internal order to GXBeam order

GXBeamCS.resample — Function

resample airfoil coordinates x, y with linear interpolation, making sure to exactly hit the points in vector xbreak (e.g., [0, 0.2, 0.4, 0.7, 1.0]). ds defines the target size between coordinates (rounded to nearnest number of grid points). assumes x is already sorted (and thus represents either upper or lower half or airfoil coordinates) points in xbreak do not have to exist in the original x vector alternatively you can define nseg, a vector of size: length(xbreak) - 1, defining how many points to use in each interval (e.g., [15, 20, 40, 30]) If nseg is provided then ds is ignored. During optimization either keep mesh fixed, or if changes are needed, then nseg should be used rather than ds to avoid discrete changes.

GXBeamCS.rotate_element_to_beam — Method

rotate_element_to_beam(Q, c, s)

Rotate stiffness matrix Q by element orientation beta where c = cos(beta) and s = sin(beta)

GXBeamCS.rotate_ply_to_element — Method

rotate_ply_to_element(Q, theta)

Rotate stiffness matrix Q by ply angle theta

GXBeamCS.rotate_stress_and_strains — Method

from laminate to beam

GXBeamCS.stiffness — Method

stiffness(material)

Construct constitutive matrix Q for the specified material (with the internal ordering).

GXBeamCS.strains — Method

strains for a thin laminate with forces: N1, N1, N12, M1, M2, M12

GXBeamCS.strains_and_stresses — Method

strains_and_stresses(F, M, fem::FEM)

Compute stresses and strains at each element in cross section.

Arguments

F::Vector(3): force at this cross section in x, y, z directions

M::Vector(3): moment at this cross section in x, y, z directions

fem::FEM: needs to reuse data from the compliance solve (thus must initialize cache and pass it to both compliance and this function)

Returns

strain_b::Vector(6, ne): strains in beam coordinate system for each element. order: xx, yy, zz, xy, xz, yz Note: this order (as well as those below) corresponds to the local beam reference frame if gxbeam_order is set to true.

stress_b::Vector(6, ne): stresses in beam coordinate system for each element. order: xx, yy, zz, xy, xz, yz

strain_p::Vector(6, ne): strains in ply coordinate system for each element. order: 11, 22, 33, 12, 13, 23

stress_p::Vector(6, ne): stresses in ply coordinate system for each element. order: 11, 22, 33, 12, 13, 23

GXBeamCS.strains_and_stresses — Method

strains_and_stresses(F, M, ::CompositeSectionAnalysis)

Compute stresses and strains across the cross section (locations depend on the method). Uses GXBeam local beam axis (e.g. axial stresses would correspond to the xx direction, or first index).

Arguments

F::Vector(3): force at this cross section in x, y, z directions. Using GXBeam local beam axis.

M::Vector(3): moment at this cross section in x, y, z directions. Using GXBeam local beam axis.

Returns

strain_b::Vector(6, nloc): strains in beam coordinate system for each element. order: xx, yy, zz, xy, xz, yz the meaning of the locations varies with the different methods (e.g., elements vs edges of plys)

stress_b::Vector(6, nloc): stresses in beam coordinate system for each element. order: xx, yy, zz, xy, xz, yz

strain_p::Vector(6, nloc): strains in ply coordinate system for each element. order: 11, 22, 33, 12, 13, 23

stress_p::Vector(6, nloc): stresses in ply coordinate system for each element. order: 11, 22, 33, 12, 13, 23

GXBeamCS.stresses — Method

stress for a thin laminate given stresses.

GXBeamCS.tangential — Method

determine tangential direction for each point on airfoil (as a normal vector with magnitudes tx and ty). keyword signifies whether this is the upper or lower surface

GXBeamCS.te_inner_intersection — Method

given curves for inner surfaces, find where they intersect (if they intersect), for handling trailing-edge mesh returns index of upper and lower surface where the trailing-edge-mesh should start (before intersection). also returns the x, y location of intersection.

GXBeamCS.tsai_wu — Method

tsai_wu(stress_p, fem::FEM)

Tsai Wu failure criteria

Arguments

stress_p::vector(6, ne): stresses in ply coordinate system

fem::FEM: finite element object

Returns

failure::vector(ne): tsai-wu failure criteria for each element. fails if >= 1

GXBeamCS.tsai_wu — Method

tsai_wu(stress_p, ::CompositeSectionAnalysis)

Tsai Wu failure criteria

Arguments

stress_p::vector(6, nloc): stresses in ply coordinate system

Returns

failure::vector(nloc): tsai-wu failure criteria at each location. fails if >= 1

GXBeamCS.web_intersections — Method

determine where web intersects the inner surface curves. creates a new discretization where points are added to line up with the vertical mesh in the web. also returns the x vector index where the web mesh should start

GXBeamCS.zspacing — Method

compute the z locations and overall height for each lamina in the laminate.

GXBeamCS.zspacingdouble_fromz — Method

Same z locations, but doubles them up on the interior so that strain can be computed at either side of the ply.